Split default themes
The "new" default is intended to be more inline with the made-for-print style of traditional unstyled HTML, but still includes a dark palette.
The "Fancy" one is intended to be more colorful and smooth. Where print finds shaded panels wasteful, they look nice on the web and are considered a standard.